pyproject.toml
[project]
name = "comp-vis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache>=0.12.3",
    "bcrypt>=4.3.0",
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "locust>=2.33.2",
    "loguru>=0.7.3",
    "mediapipe>=0.10.21",
    "mkdocs-material>=9.6.9",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.2",
    "ray[serve]>=2.43.0",
    "ruff>=0.9.9",
    "starlette>=0.46.1",
    "streamlit>=1.43.0",
    "streamlit-option-menu>=0.4.0",
    "watchfiles>=1.0.4",
    "zmq>=0.0.0",
]

[tool.ruff]
line-length = 200

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff.lint.pylint]
# This was required as some frontend functions need to be over the threshold statements length. Splitting them causes UI to fail
max-statements = 70
